backend/requirements.txt
# V-Legal Bot - RAG Chatbot Pháp Luật
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# RAG & LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.10
google-generativeai>=0.4.0

# Embeddings
sentence-transformers>=2.5.0

# Vector Database
chromadb>=0.4.20

# Document Processing
python-docx>=1.1.0
python-pptx>=0.6.20

# Evaluation & Monitoring
evidently>=0.4.20

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
tenacity>=8.2.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0